Split groups of function into corresponding files add new files reminder.py, quotes.py update list of cogs in Main.py
