+++ Main.py
@@ -18,10 +18,12 @@ from config import parser
 
 # List the extensions (modules) that should be loaded on startup.
 startup = [
-    "cogs.db",
+    "cogs.reminder",
     "cogs.memes",
     "cogs.helpers",
-    "cogs.mod"
+    "cogs.mod",
+    "cogs.score",
+    "cogs.quotes"
 ]
 
 bot = commands.Bot(command_prefix='?')
@@ -95,82 +97,6 @@ async def update(ctx):
     '''
     await ctx.send('https://streamable.com/c7s2o')
     os.system('git pull')
-
-
-@bot.event
-async def on_raw_reaction_add(payload):
-    # Check for Martlet emoji + upmartletting yourself
-    channel = bot.get_channel(payload.channel_id)
-    message = await channel.get_message(payload.message_id)
-    user = bot.get_user(payload.user_id)
-    emoji = payload.emoji
-
-    if user == message.author:
-        return
-
-    if emoji.name == 'upmartlet':
-        score = 1
-    elif emoji.name == 'downmartlet':
-        score = -1
-    else:
-        return
-
-    conn = sqlite3.connect(bot.config.db_path)
-    c = conn.cursor()
-    # uncomment to enable sqlite3 debugging
-    # conn.set_trace_callback(print)
-
-    t = (message.author.id,)
-    if not c.execute('SELECT * FROM Members WHERE ID=?', t).fetchall():
-        t = (message.author.id, message.author.display_name, score)
-        c.execute('INSERT INTO Members VALUES (?,?,?)', t)
-        conn.commit()
-        conn.close()
-    else:
-        if score == 1:
-            c.execute('UPDATE Members SET Score=Score+1 WHERE ID=?', t)
-        else:
-            c.execute('UPDATE Members SET Score=Score-1 WHERE ID=?', t)
-        conn.commit()
-        conn.close()
-
-
-@bot.event
-async def on_raw_reaction_remove(payload):
-    # Check for Martlet emoji + upmartletting yourself
-    """Does the opposite thing when a user up/downmarlets a message
-    """
-    channel = bot.get_channel(payload.channel_id)
-    message = await channel.get_message(payload.message_id)
-    user = bot.get_user(payload.user_id)
-    emoji = payload.emoji
-
-    if user == message.author:
-        return
-
-    if emoji.name == 'upmartlet':
-        score = -1
-    elif emoji.name == 'downmartlet':
-        score = 1
-    else:
-        return
-
-    conn = sqlite3.connect(bot.config.db_path)
-    c = conn.cursor()
-
-    t = (message.author.id,)
-    if not c.execute('SELECT * FROM Members WHERE ID=?', t).fetchall():
-        t = (message.author.id, message.author.display_name, score)
-        c.execute('INSERT INTO Members VALUES (?,?,?)', t)
-        conn.commit()
-        conn.close()
-    else:
-        if score == 1:
-            c.execute('UPDATE Members SET Score=Score+1 WHERE ID=?', t)
-        else:
-            c.execute('UPDATE Members SET Score=Score-1 WHERE ID=?', t)
-        conn.commit()
-        conn.close()
 
 
 @bot.event
